flask/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.11
plotly==5.17.0
python-dotenv==1.0.0
pandas
